+++ metafunctions/util.py
@@ -3,11 +3,32 @@ Utility functions for use in function pipelines.
 '''
 import sys
 import re
+import functools
 
 import colors
 
-from metafunctions.decorators import node
 from metafunctions.core import MetaFunction
+from metafunctions.core import SimpleFunction
+
+
+def node(_func=None, *, bind=False):
+    '''Turn the decorated function into a MetaFunction.
+
+    Args:
+        _func: Internal use. This will be the decorated function if node is used as a decorator with no params.
+        bind: If True, the MetaFunction object is passed to the function as its first parameter.
+
+    Usage:
+
+    @node(bind=True)
+    def f(metafunc, x):
+       <do something cool>
+    '''
+    if not _func:
+        def decorator(function):
+            return SimpleFunction(function, bind)
+        return decorator
+    return SimpleFunction(_func, bind)
 
 
 def store(key):
@@ -31,7 +52,7 @@ def recall(key, from_meta:MetaFunction=None):
     return f
 
 
-def system_supports_color():
+def _system_supports_color():
     """
     Returns True if the running system's terminal supports color, and False
     otherwise.
@@ -46,7 +67,7 @@ def system_supports_color():
     return True
 
 
-def highlight_current_function(meta, color=colors.red, use_color=system_supports_color()):
+def highlight_current_function(meta, color=colors.red, use_color=_system_supports_color()):
     '''Return a formatted string showing the location of the currently active function in meta.
 
     Consider this a 'you are here' when called from within a function pipeline.
@@ -67,4 +88,24 @@ def highlight_current_function(meta, color=colors.red, use_color=system_supports
     highlighted_string = re.sub(regex, fr'\1{highlighted_name}', str(meta))
     return highlighted_string
 
+
+def raise_with_location(method):
+    '''
+    Wrap the given method in an exception handler that intercepts exceptions to add information
+    about where in the function pipeline they occured. Note that for this to work, the outer
+    function in the function must be present as kwargs['meta'].
+
+    SimpleFunction applies this decorator by default.
+    '''
+    @functools.wraps(method)
+    def new_call(self, *args, **kwargs):
+        try:
+            return method(self, *args, **kwargs)
+        except Exception as e:
+            meta = kwargs.get('meta')
+            detailed_message = str(e)
+            if meta:
+                detailed_message = f"{str(e)} \n\n Occured in the following function: {highlight_current_function(meta)}"
+            raise type(e)(detailed_message).with_traceback(e.__traceback__)
+    return new_call
